requirements.txt
websocket-client
pypresence